actualizaciones

+++ src/dao/mongo.dao.js
@@ -1,4 +1,4 @@
-import mongoose from 'mongoose';
+import mongoose, { mongo } from 'mongoose';
 
 export default class MongoDao {
   constructor(collection, schema) {
@@ -7,10 +7,10 @@ export default class MongoDao {
 
   async getAll() {
     try {
-      const file = await this.model.find({});
+      const file = await this.model.find({}).lean().exec();
       return file;
     } catch (error) {
-      return error;
+      throw error;
     }
   }
 
@@ -21,34 +21,52 @@ export default class MongoDao {
       await newElement.save();
       return newElement;
     } catch (error) {
-      return error;
+      throw error;
     }
   }
 
   async getById(id) {
     try {
-      const file = await this.model.findById(id);
+      if (!mongoose.isValidObjectId(id)) throw { status: 400, error: 'El id indicado no es válido' };
+
+      const file = await this.model.findById(id).lean().exec();
       return file;
     } catch (error) {
-      return error;
+      throw error;
     }
   }
 
   async updateById(id, newData) {
     try {
+      if (!mongoose.isValidObjectId(id)) throw { status: 400, error: 'El id indicado no es válido' };
+
       newData.timestamp = Date.now();
-      let updatedElement = await this.model.findOneAndUpdate({ _id: id }, { ...newData }, { new: true });
+      let updatedElement = await this.model
+        .findOneAndUpdate({ _id: id }, { ...newData }, { new: true })
+        .lean()
+        .exec();
       return updatedElement;
     } catch (error) {
-      return error;
+      throw error;
     }
   }
 
   async deleteById(id) {
     try {
+      if (!mongoose.isValidObjectId(id)) throw { status: 400, error: 'El id indicado no es válido' };
+
       await this.model.findByIdAndDelete(id);
     } catch (error) {
-      return error;
+      throw error;
+    }
+  }
+
+  async getByEntity(entity) {
+    try {
+      const file = this.model.findOne({ entity: entity }).lean().exec();
+      return file;
+    } catch (error) {
+      throw error;
     }
   }
 }
